Revert "Convert v2 flag array from int32 to uint8"
This reverts commit 1db10f3b8f74fdf0bd5a6b30ce7d16bf317482c8.
Using uint8 instead of int32 in v2 seems to cause (minor)
coalescing issues, compared to v4 which has None due to
the different memory layout. Revert, rather than remove the change
so as to keep a record of it.

+++ montblanc/impl/biro/v2/BiroSolver.py
@@ -149,7 +149,7 @@ A = [
         default=1,
         test=lambda slvr, ary: (rary(ary)-0.5)*1e-2),
 
-    ary_dict('flag', (4, 'ntime', 'nbl', 'nchan'), np.uint8,
+    ary_dict('flag', (4, 'ntime', 'nbl', 'nchan'), np.int32,
         default=0,
         test=lambda slvr, ary: np.random.random_integers(
             0, 1, size=ary.shape)),
